YQ-2735 Hopping on analytic (v2) (#1029)
