+++ ydb/library/yql/dq/opt/ya.make
@@ -15,6 +15,7 @@ SRCS(
     dq_opt.cpp
     dq_opt_build.cpp
     dq_opt_join.cpp
+    dq_opt_hopping.cpp
     dq_opt_log.cpp
     dq_opt_peephole.cpp
     dq_opt_phy_finalizing.cpp
