+++ ydb/library/yql/providers/dq/common/ya.make
@@ -8,6 +8,7 @@ PEERDIR(
     ydb/library/yql/utils/log
     ydb/library/yql/dq/actors
     ydb/library/yql/dq/proto
+    ydb/library/yql/dq/integration
 )
 
 GENERATE_ENUM_SERIALIZATION(yql_dq_settings.h)
